fixed session randomly disappearing when using outlook oauth

+++ app.js
@@ -14,17 +14,29 @@ var moment = require("moment");
 
 const AdminView = require("./AdminView");
 
-var session = require('express-session')({
+var session = require('express-session');
+var MongoDBStore = require('connect-mongodb-session')(session);
+
+var store = new MongoDBStore(
+    {
+        uri: 'mongodb://localhost:27017/control',
+        collection: 'mySessions'
+    });
+
+
+var session2 = require('express-session')({
     secret: 'AdminView',
     resave: false,
     saveUninitialized: true,
     cookie: {
-        httpOnly: true,
+        httpOnly: false,
         secure: false,
-        sameSite: true
-    }
+        sameSite: false
+    },
+	store: store
 });
-app.use(session);
+app.use(session2);
+
 var sharedsession = require("express-socket.io-session");
 ios.use(sharedsession(session, {
     autoSave:true
@@ -140,7 +152,7 @@ ios.on('connection', function(socket){
         	return callback({success:false, message: "Use different username."});
 		}
 		if (socket.handshake.session.username) {
-            if (room && room.inviteOnly && !socket.handshake.session.isInstructor && !socket.handshake.session.utorid) {
+            if (room && room.inviteOnly && !socket.handshake.session.isInstructor && !socket.handshake.session.utorid && !socket.handshake.session.isAdmin) {
                 destroySession();
                 return callback({success: false, message: "Room is invite only."});
             }
